move cookies to wrappers and use component from lib
