+++ layouts/wrappers/cookies/styles.scss
@@ -1,0 +1,8 @@
+@import '~styles/settings.scss';
+
+.c-cookies {
+  position: fixed;
+  width: 100%;
+  bottom: 0;
+  z-index: 99999;
+}
